# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Load Balancer Clients API
  description: API для управления клиентами и рейтлимитами в балансировщике нагрузки
  version: 1.0.0
  contact:
    name: Ivan Kurochkin
    email: kurochkinivan@example.com

servers:
  - url: http://localhost:8080
    description: Локальный сервер

tags:
  - name: clients
    description: Управление клиентами и рейтлимитами

components:
  schemas:
    Client:
      type: object
      required:
        - id
        - ip_address
        - capacity
        - rate_per_second
      properties:
        id:
          type: integer
          format: int64
          description: Уникальный идентификатор клиента
        ip_address:
          type: string
          description: IP-адрес клиента
        capacity:
          type: integer
          format: int32
          description: Максимальное количество токенов
        rate_per_second:
          type: integer
          format: int32
          description: Скорость пополнения токенов (в токенах в секунду)

    CreateClientRequest:
      type: object
      required:
        - ip_address
        - capacity
        - rate_per_second
      properties:
        ip_address:
          type: string
          example: "192.168.0.1"
          description: IP-адрес клиента
        capacity:
          type: integer
          format: int32
          example: 100
          description: Максимальное количество токенов
        rate_per_second:
          type: integer
          format: int32
          example: 10
          description: Скорость пополнения токенов

    UpdateClientRequest:
      type: object
      required:
        - capacity
        - rate_per_second
      properties:
        capacity:
          type: integer
          format: int32
          example: 100
          description: Максимальное количество токенов
        rate_per_second:
          type: integer
          format: int32
          example: 10
          description: Скорость пополнения токенов

    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
          example: 409
        message:
          type: string
          example: "client already exists"

paths:
  /v1/api/clients/:
    get:
      tags:
        - clients
      summary: Получить список всех клиентов
      description: Возвращает массив всех клиентов с их лимитами
      responses:
        '200':
          description: Список клиентов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Client'
        '500':
          description: Ошибка на стороне сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - clients
      summary: Создать нового клиента
      description: Добавляет нового клиента с указанными параметрами рейтлимита
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateClientRequest'
      responses:
        '201':
          description: Клиент успешно создан
        '400':
          description: Невалидный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Клиент уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Ошибка на стороне сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /v1/api/clients/{ip_address}:
    put:
      tags:
        - clients
      summary: Обновить параметры клиента по IP-адресу
      description: Обновляет параметры рейтлимита клиента по его IP-адресу
      parameters:
        - name: ip_address
          in: path
          required: true
          description: IP-адрес клиента
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateClientRequest'
      responses:
        '200':
          description: Клиент успешно обновлен
        '400':
          description: Невалидный запрос или не указан IP-адрес
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Клиент не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Ошибка на стороне сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - clients
      summary: Удалить клиента по IP-адресу
      description: Удаляет клиента из системы по его IP-адресу
      parameters:
        - name: ip_address
          in: path
          required: true
          description: IP-адрес клиента
          schema:
            type: string
      responses:
        '204':
          description: Клиент успешно удалён
        '400':
          description: Не указан IP-адрес
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Клиент не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Ошибка на стороне сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'